lvm_full_driving/models/driving_model.py
from sklearn.ensemble import RandomForestClassifier
import numpy as np
import tensorflow as tf
from tensorflow import keras

class DrivingModel:
    def __init__(self):
        self.model = keras.Sequential([
            keras.layers.Flatten(input_shape=(image_height, image_width, 3)),  # Adjust input shape as needed
            keras.layers.Dense(128, activation='relu'),
            keras.layers.Dense(64, activation='relu'),
            keras.layers.Dense(3)  # Output layer for throttle, steer, and brake
        ])
        self.model.compile(optimizer='adam', loss='mse')

    def train(self, X, y):
        self.model.fit(X, y, epochs=10)  # Adjust epochs and batch size as needed

    def predict(self, X):
        return self.model.predict(X)