lvm_full_driving/scripts/reinforcement_learn.py
import gym
from stable_baselines3 import PPO
from stable_baselines3.common.envs import DummyVecEnv
from stable_baselines3.common.evaluation import evaluate_policy

# Custom CARLA environment (assuming we have a wrapper for CARLA)
from carla_env import CarlaEnv

# Create the environment
env = DummyVecEnv([lambda: CarlaEnv()])

# Load the pre-trained PPO model
model = PPO.load("path/to/pretrained/ppo/model")

# Fine-tune with reinforcement learning
model.learn(total_timesteps=100000)

# Evaluate the policy
mean_reward, std_reward = evaluate_policy(model, env, n_eval_episodes=10)
print(f"Mean reward: {mean_reward}, Std reward: {std_reward}")

# Save the fine-tuned model
model.save("ppo_carla_finetuned")