lvm_full_driving/requirements.txt
carla
openai
numpy
matplotlib
Pillow
requests
pandas
scikit-learn
tensorflow
stable-baselines3
gym
transformers

datasets
opencv-python
open3d
matplotlib
python-dotenv